\name{zTree-package}
\alias{zTree-package}
\alias{zTree}
\docType{package}
\title{
\packageTitle{zTree}
}
\description{
\packageDescription{zTree}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{zTree}
\packageIndices{zTree}
}
\author{
\packageAuthor{zTree}
Maintainer: \packageMaintainer{zTree}
}
\references{Urs Fischbacher. "z-Tree: Zurich Toolbox for Ready-made Economic Experiments."
  Experimental Economics (2) 171-178, 2007.
}
\keyword{package}
\examples{
\dontrun{
zTT <- zTreeTables(c("080712_1412.xls","080712_1818.xls"))
}}
